Recover game-domain discriminants and scalar boundaries (#303)
Advance imperialism-decomp-1uj.99 with evidence-backed domain enums, scalar representation boundaries, bool auditing, and the recovered diplomacy/news event model.

+++ include/game/TTraderAmtBar.h
@@ -10,7 +10,7 @@ public:
   // ~TTraderAmtBar is compiler-generated (implicit virtual dtor).
   DECLARE_DYNCREATE(TTraderAmtBar)
   void DoPostCreate(int arg) override;
-  int ApplyMoveClamp(int baseValue, int requestedValue) override;
+  short ApplyMoveClamp(int baseValue, int requestedValue) override;
   void RenderPrimarySurfaceOverlayPanelWithClipCache() override;
 
   void UpdateFromScaleOrRatio(int scaleValue, int ratioValue);
